Create Computer choice and User Input
The computer needs to choose a random tools to play the game.
The user needs to input a tool to play the game.
Some users may input the tool name without an uppercased first letter,
so the text will be corrected to appear exactly like it's shown in the
variables.
When the computer and the player make the same choice, it's draw, and a
new game will start **to be implemented in the next push